Error/Tests: Replace DummyLogger with LoggerMock.
